split object_finder to own file
